New option ('-a','--add_keywords')
The new option allows one to specify a target keyword (such as a
function name). Only function names that lead to that target node
are included in the final graph. For instance, if the target keyword
`d` is chosen, then the following graph:

      n --> m ----+
                  V
a --> b --> c --> d --> e
                        ^
                        |
x --> y --> z --> w ----+

is pruned and the following subgraph is output to the PDF document:

      n --> m ----+
                  V
a --> b --> c --> d